+++ CHANGELOG.md
@@ -13,3 +13,10 @@ All notable changes to this project will be documented in this file.
 
 - Add key parameter to text_annotator function
 - Add a special button to show all the annotations together
+
+## [0.0.3] - 2025-07-15
+
+### Added
+
+- Add a popup view to show more information about the annotation
+- Add the possibility to customize the popup metadata and the popup delay